Fix picom bug when running autorandr
When plugging in a monitor and changing the autorandr profile,
picom used to put the notebook monitor under a black screen. This
is solved by restarting picom.

This commit changes the `autostart.sh` script to always restart
picom and adds a switch to autorandr to automatically run the
`autostart.sh` script after changing autorandr profiles.